+++ lib/redis/daily_set.rb
@@ -1,18 +1,10 @@
 # frozen_string_literal: true
 
 require "#{File.dirname(__FILE__)}/base_set_object"
+require "#{File.dirname(__FILE__)}/recurring_at_intervals/daily"
 
 class Redis
   class DailySet < BaseSetObject
-    private
-
-    def redis_daily_field_key(date_or_time)
-      date_key = date_or_time.strftime('%Y-%m-%d')
-      [original_key, date_key].flatten.join(':')
-    end
-
-    def next_key(date, length = 1)
-      date + length
-    end
+    include Daily
   end
 end
