+++ lib/redis/hourly_set.rb
@@ -1,18 +1,10 @@
 # frozen_string_literal: true
 
 require "#{File.dirname(__FILE__)}/base_set_object"
+require "#{File.dirname(__FILE__)}/recurring_at_intervals/hourly"
 
 class Redis
   class HourlySet < BaseSetObject
-    private
-
-    def redis_daily_field_key(time)
-      time_key = time.strftime('%Y-%m-%dT%H')
-      [original_key, time_key].flatten.join(':')
-    end
-
-    def next_key(time, length = 1)
-      time + 3600 * length
-    end
+    include Hourly
   end
 end
