+++ lib/redis/monthly_set.rb
@@ -1,18 +1,10 @@
 # frozen_string_literal: true
 
 require "#{File.dirname(__FILE__)}/base_set_object"
+require "#{File.dirname(__FILE__)}/recurring_at_intervals/monthly"
 
 class Redis
   class MonthlySet < BaseSetObject
-    private
-
-    def redis_daily_field_key(date_or_time)
-      date_key = date_or_time.strftime('%Y-%m')
-      [original_key, date_key].flatten.join(':')
-    end
-
-    def next_key(date, length = 1)
-      date.next_month(length)
-    end
+    include Monthly
   end
 end
