+++ lib/redis/recurring_at_intervals/weekly.rb
@@ -1,0 +1,18 @@
+# frozen_string_literal: true
+
+class Redis
+  module RecurringAtIntervals
+    module Weekly
+      private
+
+      def redis_daily_field_key(date_or_time)
+        date_key = date_or_time.strftime('%YW%W')
+        [original_key, date_key].flatten.join(':')
+      end
+
+      def next_key(date, length = 1)
+        date + 7 * length
+      end
+    end
+  end
+end
