+++ lib/redis/weekly_counter.rb
@@ -1,18 +1,10 @@
 # frozen_string_literal: true
 
 require "#{File.dirname(__FILE__)}/base_counter_object"
+require "#{File.dirname(__FILE__)}/recurring_at_intervals/weekly"
 
 class Redis
   class WeeklyCounter < BaseCounterObject
-    private
-
-    def redis_daily_field_key(date_or_time)
-      date_key = date_or_time.strftime('%YW%W')
-      [original_key, date_key].flatten.join(':')
-    end
-
-    def next_key(date, length = 1)
-      date + 7 * length
-    end
+    include Weekly
   end
 end
